Adding Travis support

+++ build.sh
@@ -28,8 +28,7 @@
 #
 # __author__ = 'fla'
 
-virtualenv ENV
-source ENV/bin/activate
+
 mkdir /var/log/fiware-facts
 mkdir -m 777 -p target/site/cobertura
 mkdir -m 777 -p target/surefire-reports
@@ -37,24 +36,29 @@ chmod 777 /var/log/fiware-cloto
 sudo pip install -r requirements.txt
 sudo pip install -r requirements_dev.txt
 
-#INSTALLING REDIS
-wget -O redis.tar.gz http://download.redis.io/releases/redis-2.8.19.tar.gz
-sleep 12
-tar -xzvf redis.tar.gz
-cd redis-2.8.19
-make
-cd src
-./redis-server &
-cd ../..
+if [ ! $1 = "travis_build" ];
+then
+    virtualenv ENV
+    source ENV/bin/activate
+    #INSTALLING REDIS
+    wget -O redis.tar.gz http://download.redis.io/releases/redis-2.8.19.tar.gz
+    sleep 12
+    tar -xzvf redis.tar.gz
+    cd redis-2.8.19
+    make
+    cd src
+    ./redis-server &
+    cd ../..
 
-#INSTALLING RABBITMQ
-wget http://www.rabbitmq.com/releases/rabbitmq-server/v3.4.3/rabbitmq-server-3.4.3-1.noarch.rpm
-sleep 10
-yum install erlang
-rpm --import http://www.rabbitmq.com/rabbitmq-signing-key-public.asc
-yum install rabbitmq-server-3.4.3-1.noarch.rpm
-rm -rf rabbitmq-server-3.4.3-1.noarch.rpm
-sudo /sbin/service rabbitmq-server start
+    #INSTALLING RABBITMQ
+    wget http://www.rabbitmq.com/releases/rabbitmq-server/v3.4.3/rabbitmq-server-3.4.3-1.noarch.rpm
+    sleep 10
+    yum install erlang
+    rpm --import http://www.rabbitmq.com/rabbitmq-signing-key-public.asc
+    yum install rabbitmq-server-3.4.3-1.noarch.rpm
+    rm -rf rabbitmq-server-3.4.3-1.noarch.rpm
+    sudo /sbin/service rabbitmq-server start
+fi
 
 python server.py &
 export PYTHONPATH=$PWD
